# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: TI-Toolbox job server (tit.server)
  version: "1.0.0"
  description: >-
    v1 contract for the v3 Electron GUI. Every path and schema from
    `contracts/openapi.v0.yaml` is carried unchanged (the real server's `--dump-openapi` output
    must be a superset of both; CI gate `dev/contracts_check.py`); v1 adds the job engine, catalog
    CRUD, ViewSpec viewers, jailed file access, validate/plan preflight, settings, and the
    `/ws/jobs` live-job stream. Auth is unchanged from v0: the launcher passes
    `?token=<TIT_SERVER_TOKEN>` exactly once; the server mints a random session id (never the
    token), keeps it in memory, sets it as an HttpOnly, SameSite=Strict `tit_session` cookie and
    redirects to `/`. A server restart forgets every session (401 -> back to the launcher). API
    routes accept either that cookie or `Authorization: Bearer <token>` (scripts). Unauthenticated
    requests to any route other than `/api/health` and `/auth/session` get 401.
    `POST /auth/logout` forgets the session and clears the cookie.

    Config request bodies (`config` fields below) are `dict[str, Any]`, validated server-side with
    `tit.config_io.deserialize_config`. Their real shapes are the `contracts/schema.json` `$defs`
    (`SimulationConfig`, `Montage`, `FlexConfig`, `ExConfig`, `MExConfig`, `AnalyzerConfig`,
    `PreprocessConfig`, `QSIPrepConfig`, `QSIReconConfig`, `GroupComparisonConfig`,
...
    today.

...

...

...
    difference is never reported as drift.
...
servers:
...
paths:
...
  /api/system:
    get:
      tags: [system]
      summary: One system snapshot (same payload as the /ws/system stream)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/system/terminate:
    post:
      tags: [system]
      summary: Terminate a toolbox-relevant host process by pid (System screen "Terminate" action)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pid]
              properties:
                pid: { type: integer }
      responses:
        "204":
          description: terminated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: no such process
...
  /api/settings:
    get:
      tags: [settings]
      summary: Server/project settings and app-level preferences
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [settings]
      summary: Replace server/project settings
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          $ref: "#/components/responses/Unauthorized"


  # ---------------------------------------------------------------------------------- pipelines (v1)
  # A pipeline is a DAG of existing job kinds (contracts/pipeline.schema.json). Running one is
  # exactly one `submit_plan` call, so a pipeline run is ONE job group with ONE group_id -- there
  # is no pipeline executor and no new job kind. See tit/pipeline/ and docs/wiki/pipelines.md.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tit_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: missing or invalid session/token
...
  schemas:
...
    SystemSnapshot:
      type: object
      required: [ts, cpu_percent, cpu_count, mem, disk, processes]
      properties:
        ts: { type: number, description: unix seconds }
        cpu_percent: { type: number }
        cpu_count: { type: integer }
        mem:
          type: object
          required: [total, available, used, percent]
          properties:
            total: { type: integer }
            available: { type: integer }
            used: { type: integer }
            percent: { type: number }
        disk:
          type: object
          required: [total, free, percent]
          properties:
            total: { type: integer }
            free: { type: integer }
            percent: { type: number }
        processes:
          type: array
          description: toolbox-relevant processes (same keyword filter as system_monitor_tab.py)
          items:
            type: object
            required: [pid, name, cpu_percent, rss, started]
            properties:
              pid: { type: integer }
              name: { type: string }
              cmdline: { type: string }
              cpu_percent: { type: number }
              rss: { type: integer }
              started: { type: number }

    # -------------------------------------------------------------------------------- catalog (v1)
...
    Telemetry:
      type: object
      required: [consented, enabled]
      properties:
        consented: { type: boolean }
        enabled: { type: boolean }
    Settings:
      type: object
      required: [telemetry, panels, allow_unsafe_overrides, theme]
      properties:
        telemetry: { $ref: "#/components/schemas/Telemetry" }
        panels:
          type: array
          items: { type: string }
          description: enabled optional panels (Source, Cluster Permutation, NIfTI Group Average, Nilearn Visuals, Quick Notes, Subject Info)
        image_tag: { type: [string, "null"] }
        allow_unsafe_overrides: { type: boolean }
        theme: { type: string, enum: [system, light, dark] }

    # --------------------------------------------------------------- config placeholders (F1b fills)
...